backend/Her2/requirements.txt
# Wheel index with CPU-only PyTorch builds
--extra-index-url https://download.pytorch.org/whl/cpu

# Core runtime dependencies
joblib==1.3.2
matplotlib==3.10.7
numpy<2
pandas==2.3.3
Pillow==12.0.0
scikit-learn==1.2.2
streamlit==1.51.0

# Optional web endpoints
flask==2.3.3

# Surrogate Grad-CAM stack
torch==2.4.0+cpu
torchvision==0.19.0+cpu
torchaudio==2.4.0+cpu

# Image processing for safety checks
opencv-python-headless==4.10.0.84
